Fix the bug of systemWatcher can not find file
If a file is a temp file, it will be created and deleted in a very quick way, thus leads to error of system watcher. RQL will create a watcher when tmp file exists, and when checking its modify time, tmp file is already deleted. This update add a file extension name restriction to system watcher, making it check only .py, .pyt, .xlsx file, and add a try-except to systemGuardian to make sure it won't break.