Fix Controller Bug
If Dir name is changed, mimic builder will fail when try to clean the content that is built before, due to non-exist of original path.

This bug is fixed by adding check of existence of original path. If it does not exist any more, new mimic builder will be created.